irmin-pack.unix: Add an interface.
This adds an interface file (irmin_pack_unix.mli) for the
irmin-pack.unix package.

This is a major API change as it hides a lot of internal functionality
that was previously exposed (the interface in s.ml was not being used
in the interface of the package).

Many tests were using internal functionality. The required
functionality has been exposed in the `Internal` module.