irmin-pack: use one LRU for entire store
Previously, an `irmin-pack` store had one LRU per `Pack_store`,
resulting in three LRUs corresponding to the three object stores:
Commit, Node, and Content.  This commit changes an `irmin-pack` store to
only have one LRU that is shared by each object store. The motivation of
this change is to make configuring the LRU size more intuitive for users
of `irmin-pack`.